Basic Service Structure Created
