+++ nginx/nginx.conf
@@ -4,12 +4,53 @@ events { worker_connections 1024; }
 
 http {
 
+    # Redirect all HTTP traffic to HTTPS
     server {
-
         listen 80;
+        server_name mediconnect.live www.mediconnect.live;
+
+        # Redirect all HTTP requests to HTTPS
+        return 301 https://$host$request_uri;
+    }
+
+    server {
+        # Listen on port 443 for HTTPS
+        listen 443 ssl;
+        server_name mediconnect.live www.mediconnect.live;
+
+        # SSL certificate and key file locations
+        ssl_certificate /etc/letsencrypt/live/mediconnect.live/fullchain.pem;
+        ssl_certificate_key /etc/letsencrypt/live/mediconnect.live/privkey.pem;
+
+        # SSL settings (optional, for better security)
+        ssl_protocols TLSv1.2 TLSv1.3;
+        ssl_prefer_server_ciphers on;
+        ssl_ciphers "ECDHE-ECDSA-CHACHA20-POLY1305:ECDHE-RSA-CHACHA20-POLY1305:ECDHE-ECDSA-AES128-GCM-SHA256:ECDHE-RSA-AES128-GCM-SHA256:ECDHE-ECDSA-AES256-GCM-SHA384:ECDHE-RSA-AES256-GCM-SHA384";
+        ssl_session_cache shared:SSL:10m;
+        ssl_session_timeout 10m;
+
         charset utf-8;
 
         location / {
+            proxy_pass http://127.0.0.1:3000;
+            proxy_http_version 1.1;
+            proxy_set_header Upgrade $http_upgrade;
+            proxy_set_header Connection 'upgrade';
+            proxy_set_header Host $host;
+            proxy_cache_bypass $http_upgrade;
+        }
+
+        location /auth {
+            internal;
+            proxy_pass http://127.0.0.1:3000;
+            proxy_http_version 1.1;
+            proxy_set_header Upgrade $http_upgrade;
+            proxy_set_header Connection 'upgrade';
+            proxy_set_header Host $host;
+            proxy_cache_bypass $http_upgrade;
+        }
+
+        location /doctors {
             proxy_pass http://127.0.0.1:3001;
             proxy_http_version 1.1;
             proxy_set_header Upgrade $http_upgrade;
@@ -18,18 +59,10 @@ http {
             proxy_cache_bypass $http_upgrade;
         }
 
-        location ~ ^/users {
-            rewrite ^/users/(.*) /$1 break;
-            proxy_pass http://127.0.0.1:3001;
-            proxy_http_version 1.1;
-            proxy_set_header Upgrade $http_upgrade;
-            proxy_set_header Connection 'upgrade';
-            proxy_set_header Host $host;
-            proxy_cache_bypass $http_upgrade;
-        }
-
-        location /products {
-            rewrite ^/products/(.*)$ /$1 break;
+        location /patients {
+            #rewrite ^/products/(.*)$ /$1 break;
+            auth_request /auth;
+            auth_request_set $auth_status $upstream_status;
             proxy_pass http://127.0.0.1:3002;
             proxy_http_version 1.1;
             proxy_set_header Upgrade $http_upgrade;
