--- package.json
@@ -1,18 +1,0 @@
-{
-  "name": "mediconnect",
-  "version": "1.0.0",
-  "description": "MediConnect is a health portal that bridges that streamlines the medical formalities so that the doctors can focus on patients",
-  "main": "UserService.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "cors": "^2.8.5",
-    "express": "^4.21.0",
-    "express-http-proxy": "^2.1.1",
-    "nginx": "^1.0.8"
-  }
-}
